requirements.txt
pyrofork
google-generativeai
gTTS
flask
aiofiles
aiohttp
asyncio
beautifulsoup4
dnspython
ffmpeg-python
gitpython
hachoir
heroku3
httpx[http2]==0.27.0
motor
numpy
opencv-python
pillow==9.5.0
psutil
py-tgcalls==0.9.7
ntgcalls==1.1.2
pykeyboard
python-dotenv
pyyaml
requests
speedtest-cli
spotipy
tgcrypto
unidecode
#uvloop
yt-dlp
youtube-search
youtube-search-python
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib




